requirements.txt
requests==2.31.0
streamlit==1.30.0
streamlit-authenticator==0.3.1    
beautifulsoup4==4.12.2
Pillow==10.1.0
selenium==4.17.2
pandas==2.1.4